Keep release validation bound to the published npm artifact
Packed-artifact tests exercise filesystem writes, shared cwd, child processes, and loopback networking, so keep that coverage in the CI integration suite with its exact restricted permissions. Stable and prerelease publishing must also materialize the same SHA-addressed artifact validated by every runtime instead of rebuilding untested npm output.

Constraint: Semantic unit-boundary audit must not grow for new side-effecting unit tests
Constraint: Published npm packages must be byte-for-byte descendants of the runtime-tested canonical artifact
Rejected: Add filesystem effects to the unit baseline | masks the boundary violation
Rejected: Rebuild npm output in the stable release job | severs the tested-to-published provenance chain
Confidence: high
Scope-risk: moderate
Reversibility: clean
Directive: All release paths must consume npm-compatibility-${{ github.sha }} after the artifact quality gate
Tested: semantic disposition audit, CI TypeScript gate, exact restricted packed-artifact test, artifact workflow contract, registry release workflow contract, formatting, and diff check
Not-tested: full repository pre-push gate pending
