+++ .github/workflows/cicd.yml
@@ -1228,6 +1228,7 @@ jobs:
       tests-sentry-runtime-packages,
       tests-windows-localhost,
       build-binaries,
+      npm-compatibility-artifact,
       version-check,
     ]
     runs-on: ubuntu-latest
@@ -1292,11 +1293,18 @@ jobs:
           registry-url: "https://registry.npmjs.org"
           package-manager-cache: false
 
-      - name: Build and publish
+      - name: Download tested npm compatibility artifact
+        uses: actions/download-artifact@3e5f45b2cfb9172054b4087a40e8e0b5a5461e7c # v8
+        with:
+          name: npm-compatibility-${{ github.sha }}
+          path: dist/npm-compatibility
+
+      - name: Publish tested stable npm artifact
         env:
           VERSION: ${{ steps.version.outputs.version }}
         run: |
-          deno task build:npm
+          deno run --config=scripts/test.deno.json --frozen --allow-read --allow-write --allow-run=tar \
+            scripts/ci/npm-compatibility-artifact.ts materialize dist/npm-compatibility npm
           deno run --allow-read --allow-run=npm scripts/build/report-artifact-sizes.ts --npm-package npm \
             | tee -a "$GITHUB_STEP_SUMMARY"
           scripts/ci/publish-npm-packages.sh release-publish
