+++ scripts/test/runtime-inference-critical-flow.test.ts
@@ -10,12 +10,10 @@ import { parse } from "#std/yaml/parse";
 import {
   artifactClaim,
   assertListedRunFailure,
-  loadPackedArtifactDirectory,
   parseAgUiTextDeltas,
   parsePackedArtifactDirectory,
   parseRuntimeSelection,
   parseScopedResponseJson,
-  runRuntimeInferenceCriticalFlow,
   sanitizeRuntimeCriticalFlowFailureText,
   validateAnthropicRequest,
   waitForProviderCancellation,
@@ -28,7 +26,6 @@ import {
   parseCommaSeparatedFlag,
 } from "./runtime-e2e-helpers.ts";
 import { withMockFetch } from "#veryfront/testing/mock-fetch.ts";
-import { isAbsolute, relative } from "#std/path";
 
 const VALID_WIRE_MODEL = "claude-haiku-4-5-20251001";
 const VALID_KEY = "vf-runtime-critical-flow-key";
@@ -103,46 +100,6 @@ function unresolvedCancellationFields(): {
   };
 }
 
-async function sha256File(path: string): Promise<string> {
-  const bytes = await Deno.readFile(path);
-  const digest = await crypto.subtle.digest("SHA-256", bytes);
-  return Array.from(
-    new Uint8Array(digest),
-    (byte) => byte.toString(16).padStart(2, "0"),
-  ).join("");
-}
-
-async function writePackedArtifactFixture(
-  artifactDir: string,
-  rootBody: string,
-): Promise<string> {
-  const rootFile = `${artifactDir}/veryfront-0.1.0.tgz`;
-  await Deno.writeTextFile(rootFile, rootBody);
-  await Deno.writeTextFile(
-    `${artifactDir}/manifest.json`,
-    `${
-      JSON.stringify(
-        {
-          schemaVersion: 1,
-          rootPackage: "veryfront",
-          rootExtensionNames: [],
-          packages: [
-            {
-              name: "veryfront",
-              version: "0.1.0",
-              file: "veryfront-0.1.0.tgz",
-              sha256: await sha256File(rootFile),
-            },
-          ],
-        },
-        null,
-        2,
-      )
-    }\n`,
-  );
-  return rootFile;
-}
-
 describe("runtime inference critical-flow pure contract", () => {
   it("accepts the canonical packed artifact directory as an inline or separate flag", () => {
     assertEquals(
@@ -196,60 +153,15 @@ describe("runtime inference critical-flow pure contract", () => {
     );
   });
 
-  it("loads packed artifact tarballs as absolute paths before scaffold cwd changes", async () => {
-    const artifactDir = await Deno.makeTempDir({
-      dir: Deno.cwd(),
-      prefix: ".runtime-packed-artifact-",
-    });
-    try {
-      const extensionFile = `${artifactDir}/ext-bundler-esbuild-0.1.0.tgz`;
-      await writePackedArtifactFixture(
-        artifactDir,
-        "root package",
-      );
-      await Deno.writeTextFile(extensionFile, "extension package");
-      const manifest = JSON.parse(
-        await Deno.readTextFile(`${artifactDir}/manifest.json`),
-      );
-      manifest.rootExtensionNames = ["@veryfront/ext-bundler-esbuild"];
-      manifest.packages.push({
-        name: "@veryfront/ext-bundler-esbuild",
-        version: "0.1.0",
-        file: "ext-bundler-esbuild-0.1.0.tgz",
-        sha256: await sha256File(extensionFile),
-      });
-      await Deno.writeTextFile(
-        `${artifactDir}/manifest.json`,
-        `${JSON.stringify(manifest, null, 2)}\n`,
-      );
-
-      const loaded = await loadPackedArtifactDirectory(
-        relative(Deno.cwd(), artifactDir),
-      );
-
-      assert(isAbsolute(loaded.root), "Root tarball path must be absolute");
-      assert(
-        loaded.extensions.every(({ tarball }) => isAbsolute(tarball)),
-        "Extension tarball paths must be absolute",
-      );
-      assertEquals(await Deno.readTextFile(loaded.root), "root package");
-      assertEquals(
-        await Deno.readTextFile(loaded.extensions[0]?.tarball ?? ""),
-        "extension package",
-      );
-    } finally {
-      await Deno.remove(artifactDir, { recursive: true }).catch(() => {});
-    }
-  });
-
   it("redacts local paths from critical-flow command failures", () => {
+    const checkoutPath = "/workspace/veryfront-code";
     const message = sanitizeRuntimeCriticalFlowFailureText(
       [
         "deno task build:npm --packed-dir=/private/tmp/vf/artifact failed",
         "stderr: cannot read file:///private/tmp/vf/cache/mod.ts",
-        `stack: at main (${Deno.cwd()}/scripts/test/runtime-inference-critical-flow.ts:1:1)`,
+        `stack: at main (${checkoutPath}/scripts/test/runtime-inference-critical-flow.ts:1:1)`,
       ].join("\n"),
-      [Deno.cwd()],
+      [checkoutPath],
     );
 
     assertStringIncludes(
@@ -266,58 +178,9 @@ describe("runtime inference critical-flow pure contract", () => {
       "Redacted critical-flow failures must not include file URL paths",
     );
     assert(
-      !message.includes(Deno.cwd()),
+      !message.includes(checkoutPath),
       "Redacted critical-flow failures must not include the checkout path",
     );
-  });
-
-  it("redacts paths when a checksum-valid packed artifact is not a tarball", async () => {
-    const artifactDir = await Deno.makeTempDir({
-      dir: Deno.cwd(),
-      prefix: ".runtime-invalid-packed-artifact-",
-    });
-    try {
-      const tarball = await writePackedArtifactFixture(
-        artifactDir,
-        "this checksum matches, but this is not a gzipped tarball",
-      );
-
-      const error = await assertRejects(
-        () =>
-          runRuntimeInferenceCriticalFlow([
-            "--runtime=node",
-            "--skip-build",
-            "--packed-dir",
-            relative(Deno.cwd(), artifactDir),
-          ]),
-        Error,
-      );
-      const message = (error as Error).message;
-
-      assertStringIncludes(
-        message,
-        "node/packed npm consumer failed",
-        "The regression should exercise the runtime journey failure wrapper",
-      );
-      assert(
-        !message.includes(artifactDir),
-        "Invalid packed-artifact failures must not include the artifact directory",
-      );
-      assert(
-        !message.includes(tarball),
-        "Invalid packed-artifact failures must not include the canonical tarball path",
-      );
-      assert(
-        !message.includes(Deno.cwd()),
-        "Invalid packed-artifact failures must not include the checkout path",
-      );
-      assert(
-        !message.includes("file://"),
-        "Invalid packed-artifact failures must not include file URLs",
-      );
-    } finally {
-      await Deno.remove(artifactDir, { recursive: true }).catch(() => {});
-    }
   });
 
   it("selects all runtimes by default in stable order", () => {
